refactor(datepicker): move transformers in separate file
